backend/requirements.txt
flask
flask-cors
pandas
mysql-connector-python
python-dotenv
pydantic
playwright
beautifulsoup4
requests
fake-useragent
flask_sqlalchemy
flask_jwt_extended
chardet

